+++ utils/matrix_factory.h
@@ -27,12 +27,12 @@ class MatrixFactory {
                          long long seed = std::chrono::duration_cast<std::chrono::microseconds>
                              (std::chrono::system_clock::now().time_since_epoch()).count());
 
-  SquareMatrix<T> CreateRandomMatrix(size_t dim, bool symmetric = false);
-  SquareMatrix<T> CreateLowdiagMatrix(size_t dim);
+  SquareMatrix<T> CreateRandomMatrix(int32_t dim, bool symmetric = false);
+  SquareMatrix<T> CreateLowdiagMatrix(int32_t dim);
 
-  std::vector<T> CreateRandomVector(size_t length);
+  std::vector<T> CreateRandomVector(int32_t length);
 
-  static SquareMatrix<T> CreateIdentity(size_t dim);
+  static SquareMatrix<T> CreateIdentity(int32_t dim);
 
  private:
   std::mt19937 generator_;
@@ -45,11 +45,11 @@ MatrixFactory<T>::MatrixFactory(T lower_bound, T upper_bound, long long int seed
       distribution_(lower_bound, upper_bound) {}
 
 template<class T>
-SquareMatrix<T> MatrixFactory<T>::CreateRandomMatrix(size_t dim, bool symmetric) {
+SquareMatrix<T> MatrixFactory<T>::CreateRandomMatrix(int32_t dim, bool symmetric) {
   std::vector<std::vector<T>> data;
   data.reserve(dim);
 
-  for (size_t row_index = 0; row_index < dim; ++row_index) {
+  for (int32_t row_index = 0; row_index < dim; ++row_index) {
     std::vector<T> row;
     row.reserve(dim);
 
@@ -57,12 +57,12 @@ SquareMatrix<T> MatrixFactory<T>::CreateRandomMatrix(size_t dim, bool symmetric)
       row.resize(row_index, T(0));
     }
 
-    for (size_t column_index = (symmetric ? row_index : 0); column_index < dim; ++column_index) {
+    for (int32_t column_index = (symmetric ? row_index : 0); column_index < dim; ++column_index) {
       row.push_back(distribution_(generator_));
     }
 
     if (symmetric) {
-      for (size_t column_index = 0; column_index < row_index; ++column_index) {
+      for (int32_t column_index = 0; column_index < row_index; ++column_index) {
         row[column_index] = data[column_index][row_index];
       }
     }
@@ -74,15 +74,15 @@ SquareMatrix<T> MatrixFactory<T>::CreateRandomMatrix(size_t dim, bool symmetric)
 }
 
 template<class T>
-SquareMatrix<T> MatrixFactory<T>::CreateLowdiagMatrix(size_t dim) {
+SquareMatrix<T> MatrixFactory<T>::CreateLowdiagMatrix(int32_t dim) {
   std::vector<std::vector<T>> data;
   data.reserve(dim);
 
-  for (size_t row_index = 0; row_index < dim; ++row_index) {
+  for (int32_t row_index = 0; row_index < dim; ++row_index) {
     std::vector<T> row;
-    size_t length = std::min(dim, row_index + 2);
+    int32_t length = std::min(dim, row_index + 2);
     row.reserve(length);
-    for (size_t column_index = 0; column_index < length; ++column_index) {
+    for (int32_t column_index = 0; column_index < length; ++column_index) {
       row.push_back(distribution_(generator_));
     }
     row.resize(dim, T(0));
@@ -93,20 +93,20 @@ SquareMatrix<T> MatrixFactory<T>::CreateLowdiagMatrix(size_t dim) {
 }
 
 template<class T>
-std::vector<T> MatrixFactory<T>::CreateRandomVector(size_t length) {
+std::vector<T> MatrixFactory<T>::CreateRandomVector(int32_t length) {
   std::vector<T> result;
-  for (size_t index = 0; index < length; ++index) {
+  for (int32_t index = 0; index < length; ++index) {
     result.push_back(distribution_(generator_));
   }
   return result;
 }
 
 template<class T>
-SquareMatrix<T> MatrixFactory<T>::CreateIdentity(size_t dim) {
+SquareMatrix<T> MatrixFactory<T>::CreateIdentity(int32_t dim) {
   std::vector<T> row(dim, T(0));
   std::vector<std::vector<T>> data;
 
-  for (size_t row_index = 0; row_index < dim; ++row_index) {
+  for (int32_t row_index = 0; row_index < dim; ++row_index) {
     row[row_index] = T(1);
     data.push_back(row);
     row[row_index] = T(0);
